edit

+++ src/main/java/kg/attractor/job_search_java_25/dto/resumeDtos/ResumeViewDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,7 +18,7 @@ public class ResumeViewDto {
     private Long categoryId;
     private String categoryName;
     private String name;
-    private Float salary;
+    private BigDecimal salary;
     private boolean active;
     private LocalDateTime createdDate;
     private LocalDateTime updateTime;
